backend/requirements.txt
# Core web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Authentication and security
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.3
pycryptodome==3.21.0
authlib==0.2.1
requests==2.31.0
PyJWT==2.8.0

# Configuration and utilities  
python-dotenv==1.0.1
aiofiles==24.1.0

# Database drivers
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pymongo==4.9.2
motor==3.6.0

# Data validation
pydantic==2.10.4
pydantic-settings==2.7.0

# HTTP client and WebSocket
httpx==0.27.0
websockets==14.1

# AI/ML and cloud services
openai==1.55.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
firebase-admin==6.5.0
supabase==2.8.1

# Task queue and caching
redis==5.2.1
celery==5.4.0

# File processing
python-magic==0.4.27
Pillow==11.1.0
PyPDF2==3.0.1
python-docx==1.1.2

# Data science and ML
numpy==2.1.3
scikit-learn==1.5.2

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# Additional utilities
python-dateutil==2.8.2
pytz==2024.1